Add support for the fdtfile variable in uEnv.txt
Comments for v1:

When TorizonCore Builder applies a device tree to a TorizonCore
image, the variable fdtfile is defined in uEnv.txt, instead of
fdt_file already supported in OSTree.

This patch adds support in ostree to this variable, to make sure its
rollback variant (fdtfile2) is created in an update, and used in a
rollback situation.

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>

Comments for v2:

Fix memory leak and generalize the implementation to allow other fields
to be searched in uEnv.txt.

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>

Upstream-Status: Inappropriate [TorizonCore specific]